browser/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "browser"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "U.N. Owen", email = "void@some.where" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "websockets==15.0.1",
  "requests==2.32.3"
]

[project.urls]
Documentation = "https://github.com/U.N. Owen/browser#readme"
Issues = "https://github.com/U.N. Owen/browser/issues"
Source = "https://github.com/U.N. Owen/browser"

[tool.hatch.version]
path = "src/browser/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/browser tests}"

[tool.coverage.run]
source_pkgs = ["browser", "tests"]
branch = true
parallel = true
omit = [
  "src/browser/__about__.py",
]

[tool.coverage.paths]
browser = ["src/browser", "*/browser/src/browser"]
tests = ["tests", "*/browser/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
